refactor(hooks): relocate hooks to module-specific directories
