+++ src/modules/map/hooks/useMapRefSignals.ts
