+++ src/modules/map/hooks/usePOI.tsx
